+++ tests/unit/test_main.py
@@ -145,11 +145,13 @@ class TestMain(TestCase):
         evc1.is_enabled.return_value = True
         evc1.is_active.return_value = False
         evc1.lock.locked.return_value = False
+        evc1.recent_updated.return_value = False
         evc1.check_traces.return_value = True
         evc2 = MagicMock(id=2, service_level=7, creation_time=1)
         evc2.is_enabled.return_value = True
         evc2.is_active.return_value = False
         evc2.lock.locked.return_value = False
+        evc2.recent_updated.return_value = False
         evc2.check_traces.return_value = False
         self.napp.circuits = {'1': evc1, '2': evc2}
         assert self.napp.get_evcs_by_svc_level() == [evc2, evc1]
@@ -166,6 +168,7 @@ class TestMain(TestCase):
         evc1.is_enabled.return_value = True
         evc1.is_active.return_value = False
         evc1.lock.locked.return_value = False
+        evc1.recent_updated.return_value = False
         evc1.check_traces.return_value = False
         evc1.deploy.call_count = 0
         self.napp.circuits = {'1': evc1}